Разбивает файлы на модули, модули не подключается
