+++ index.html
@@ -233,7 +233,7 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
-<script src="functions.js"></script>
-<script src="js/main.js"></script>
+<!-- <script src="functions.js"></script> -->
+<script type="module" src="js/main.js"></script>
 </body>
 </html>
